dont template entire shrine.conf
attempting to template out the entire config file into env vars feels like "moving" the problem instead of solving it.  instead, lets expect a volume mount and only template out the passwords.  as for context.xml, this approach still seems worthwhile, as that file is credential-heavy, repetitive, and likely very similar from site-to-site.